clear package reorganization. Everything concerning carriage now in carriage package
