+++ ebu_tt_live/carriage/twisted.py
@@ -1,5 +1,5 @@
 
-from ebu_tt_live.node import ProducerCarriageImpl, ConsumerCarriageImpl
+from .base import ProducerCarriageImpl, ConsumerCarriageImpl
 from ebu_tt_live.bindings import CreateFromDocument
 from ebu_tt_live.strings import ERR_DECODING_XML_FAILED
 from ebu_tt_live.errors import XMLParsingFailed
